pyproject.toml
[project]
name = "ai_trending"
version = "0.1.0"
description = "每日 AI 开源项目与新闻聚合报告系统 — 基于 LangGraph 状态机编排的 AI 流水线"
readme = "README.md"
authors = [{ name = "Avin Zhang", email = "avinzhang@example.com" }]
license = { text = "MIT" }
requires-python = ">=3.10,<3.14"
keywords = ["ai", "trending", "github", "news", "langgraph", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langgraph>=1.1.3",
    "crewai==1.11.0",
    "litellm>=1.80.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ai_trending = "ai_trending.main:run"
run_with_trigger = "ai_trending.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[tool.black]
target-version = ["py310", "py311", "py312"]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["ai_trending"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
    "*.egg-info",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src/ai_trending",
    "--cov-report=term-missing",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["src/ai_trending"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "test.*",
]
disallow_untyped_defs = false

# CrewAI 的 @CrewBase 装饰器会动态注入 agents/tasks 属性，mypy 无法静态推断
# 同时 CrewOutput | CrewStreamingOutput 的 union 类型在 crewai 库中无法完全静态分析
[[tool.mypy.overrides]]
module = [
    "ai_trending.crew.github_trending.keyword_planning.crew",
    "ai_trending.crew.github_trending.trend_ranking.crew",
    "ai_trending.crew.new_collect.crew",
    "ai_trending.crew.trend_scoring.crew",
    "ai_trending.crew.report_writing.crew",
    "ai_trending.crew.github_trending.crew",
]
disallow_untyped_defs = false
disallow_untyped_decorators = false
warn_return_any = false
disable_error_code = ["attr-defined", "union-attr", "arg-type", "no-any-return", "call-arg"]

# 以下模块存在部分函数缺少类型注解，逐步补全中
[[tool.mypy.overrides]]
module = [
    "ai_trending.logger",
    "ai_trending.metrics",
    "ai_trending.config",
    "ai_trending.llm_client",
    "ai_trending.main",
    "ai_trending.retry",
    "ai_trending.crew.util.dedup_cache",
    "ai_trending.crew.github_trending.utils",
    "ai_trending.crew.github_trending.ranker",
    "ai_trending.crew.github_trending.fetchers",
    "ai_trending.tools.wechat_publish_tool",
    "ai_trending.tools.github_trending_tool",
]
disallow_untyped_defs = false
warn_return_any = false
warn_unused_ignores = false
disable_error_code = ["no-any-return", "operator", "arg-type"]

# LangGraph 的 StateGraph 泛型类型推断与 mypy 存在兼容性问题
[[tool.mypy.overrides]]
module = [
    "ai_trending.graph",
    "ai_trending.nodes",
]
disallow_untyped_defs = false
warn_return_any = false
disable_error_code = ["type-var", "return-value", "attr-defined", "arg-type"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B009",  # do not call getattr with a constant attribute
    "B010",  # do not call setattr with a constant attribute
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
]